DS-3150 Restructured packages
