+++ dspace-api/src/main/java/org/dspace/importer/external/exception/package-info.java
@@ -5,9 +5,9 @@
  *
  * http://www.dspace.org/license/
  */
+
 /**
- * Pubmed specific implementations of any metadatamapping services ({@link org.dspace.importer.external.metadatamapping.service})
- * @author Roeland Dillen (roeland at atmire dot com)
+ * External importer Exception classes + possible handlers
  * @author Jonas Van Goolen (jonas at atmire dot com)
  */
-package org.dspace.importer.external.pubmed.metadatamapping.service;
+package org.dspace.importer.external.exception;
