+++ dspace-api/src/main/java/org/dspace/importer/external/pubmed/metadatamapping/contributor/PubmedDateMetadatumContributor.java
@@ -6,7 +6,7 @@
  * http://www.dspace.org/license/
  */
 
-package org.dspace.importer.external.pubmed.metadatamapping;
+package org.dspace.importer.external.pubmed.metadatamapping.contributor;
 
 import org.apache.log4j.Logger;
 import org.dspace.content.DCDate;
