+++ dspace-api/src/main/java/org/dspace/importer/external/pubmed/package-info.java
@@ -7,7 +7,7 @@
  */
 
 /**
- * Pubmed implementation of {@link org.dspace.importer.external.service.other.MetadataSource}
+ * Pubmed implementation of {@link org.dspace.importer.external.service.components.MetadataSource}
  * Based on metadata support in {@link org.dspace.importer.external.metadatamapping}
  * @author Roeland Dillen (roeland at atmire dot com)
  */
