--- dspace-api/src/main/java/org/dspace/importer/external/pubmed/service/other/package-info.java
@@ -1,13 +1,0 @@
-/**
- * The contents of this file are subject to the license and copyright
- * detailed in the LICENSE and NOTICE files at the root of the source
- * tree and available online at
- *
- * http://www.dspace.org/license/
- */
-
-/**
- * Pubmed specific implementations of the generalised {@link org.dspace.importer.external.service.other} classes
- * @author Roeland Dillen (roeland at atmire dot com)
- */
-package org.dspace.importer.external.pubmed.service.other;
