+++ dspace-api/src/main/java/org/dspace/importer/external/pubmed/service/package-info.java
@@ -7,8 +7,7 @@
  */
 
 /**
- * Implementations of {@link org.dspace.importer.external.metadatamapping.service.MetadataProcessorService}
- * @author Roeland Dillen (roeland at atmire dot com)
+ * Pubmed implementation of specific {@link org.dspace.importer.external.service} classes
  * @author Jonas Van Goolen (jonas at atmire dot com)
  */
-package org.dspace.importer.external.metadatamapping.processor;
+package org.dspace.importer.external.pubmed.service;
