+++ dspace-api/src/main/java/org/dspace/submit/step/XMLUIStartSubmissionLookupStep.java
@@ -17,7 +17,7 @@ import org.dspace.content.MetadataValue;
 import org.dspace.content.factory.ContentServiceFactory;
 import org.dspace.content.service.ItemService;
 import org.dspace.core.Context;
-import org.dspace.importer.external.MetadataSourceException;
+import org.dspace.importer.external.exception.MetadataSourceException;
 import org.dspace.importer.external.datamodel.ImportRecord;
 import org.dspace.importer.external.metadatamapping.MetadatumDTO;
 import org.dspace.importer.external.service.ImportService;
